requirements.txt
discord.py
yt-dlp
PyNaCl
asyncio
ffmpeg-python
googlesearch-python
tawpy
pyjokes
beautifulsoup4